validacion al momento de reproducir una cancion que no existe

+++ src/app/screens/Music/index.tsx
@@ -3,6 +3,7 @@ import {Text, Image, View} from 'react-native';
 import {connect} from 'react-redux';
 import {destroy, getQueue} from 'react-native-track-player';
 import AsyncStorage from '@react-native-community/async-storage';
+import fs from 'react-native-fs';
 
 import {IProps} from './interfaces/Props';
 import {IState} from './interfaces/State';
@@ -22,6 +23,8 @@ import {Progress} from './components/Progress';
 import {isPlay} from '../../../utils/isPlay';
 import share from '../../../utils/share';
 import {HeaderMusic} from './components/HeaderMusic';
+import {MSong} from 'src/app/models/song.model';
+import {ShowToast} from '../../../utils/toast';
 
 class Music extends Component<IProps, IState> {
   constructor(props: IProps) {
@@ -36,7 +39,14 @@ class Music extends Component<IProps, IState> {
       musicReducer,
     } = this.props;
     // @ts-ignore
-    const item = params.item;
+    const item: MSong = params.item;
+
+    // evalua si la cancion existe en el telefono
+    if (!(await fs.exists(item.path))) {
+      ShowToast('Esta cancion no existe en el sistema');
+      this.props.navigation.goBack();
+      return;
+    }
 
     // guarda la ultima cancion reproducida
     AsyncStorage.setItem('@LastMusic', JSON.stringify(item));
